EDIT: remove baseRequestEntity and only use requestEntity
